mysite/requirements.txt
Django==2.1
boto3>=1.14.20
click>=7.1.2
PyYAML>=5.3.1
requests>=2.24.0
#zappa>=0.51.0

